# src/modules/providers/implementations/scaleway/openapi/domain.yml
openapi: 3.1.0
info:
  title: Domains and DNS API
  description: |-
    The Domains and DNS API documentation allows you to configure and manage your domains' DNS zones and records. You can also use dynamic records to optimize and easily use your infrastructure as code.
    Read our [reference content documentation](https://www.scaleway.com/en/docs/network/domains-and-dns/reference-content/) for more information about domains.

    Refer to the [Domains and DNS Registrar API documentation](https://www.scaleway.com/en/developers/api/domains-and-dns/registrar-api/) to find out how to buy,transfer and manage your domains and contacts (DNSSEC included).


    (switchcolumn)
    (switchcolumn)
    ## Concepts

    Refer to our [dedicated concepts page](https://www.scaleway.com/en/docs/network/domains-and-dns/concepts/) to find definitions of the different terms referring to Scaleway Domains and DNS.

    (switchcolumn)
    (switchcolumn)
    ## Quickstart

    1. **Configure your environment variables.**

        <Message type="note">
        This is an optional step that seeks to simplify your usage of the API.
        </Message>

        ```bash
        export SCW_ACCESS_KEY="<API access key>"
        export SCW_SECRET_KEY="<API secret key>"
        export SCW_PROJECT_ID="<Scaleway Project ID>"
        ```
    2. **Register an external domain.**

        Run the following command to register your domain:

        ```bash
...
          -X POST \
...
          -d '{
...
        }'
...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
components:
  schemas:
...
    google.protobuf.StringValue:
      type: string
      nullable: true
...
    scaleway.domain.v2beta1.DNSZone:
      type: object
      properties:
        domain:
          type: string
        subdomain:
          type: string
        ns:
          type: array
          items:
            type: string
        ns_default:
          type: array
          items:
            type: string
        ns_master:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/scaleway.domain.v2beta1.DNSZone.Status'
        message:
          $ref: '#/components/schemas/google.protobuf.StringValue'
        updated_at:
          type: string
          description: (RFC 3339 format)
          format: date-time
          example: "2022-03-22T12:34:56.123456Z"
          nullable: true
        project_id:
          type: string
        linked_products:
          type: array
          items:
            $ref: '#/components/schemas/scaleway.domain.v2beta1.LinkedProduct'
      x-properties-order:
      - domain
      - subdomain
      - ns
      - ns_default
      - ns_master
      - status
      - message
      - updated_at
      - project_id
      - linked_products
    scaleway.domain.v2beta1.DNSZone.Status:
      type: string
      enum:
      - unknown
      - active
      - pending
      - error
      - locked
      default: unknown
...
    scaleway.domain.v2beta1.DeleteDNSZoneResponse:
      type: object
...
    scaleway.domain.v2beta1.LinkedProduct:
      type: string
      enum:
      - unknown_product
      - vpc
      default: unknown_product
...
  securitySchemes:
    scaleway:
      in: header
      name: X-Auth-Token
      type: apiKey
paths:
...
  /domain/v2beta1/dns-zones/{dns_zone}:
    patch:
      tags:
      - DNS Zones
      operationId: UpdateDNSZone
      summary: Update a DNS zone
      description: Update the name and/or the Organizations for a DNS zone.
      parameters:
      - in: path
        name: dns_zone
        description: DNS zone to update.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scaleway.domain.v2beta1.DNSZone'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_dns_zone:
                  type: string
                  description: Name of the new DNS zone to create.
                  nullable: true
                project_id:
                  type: string
                  description: Project ID in which to create the new DNS zone.
              required:
              - new_dns_zone
              - project_id
              x-properties-order:
              - new_dns_zone
              - project_id
      security:
      - scaleway: []
      x-codeSamples:
      - lang: cURL
        source: |-
          curl -X PATCH \
            -H "X-Auth-Token: $SCW_SECRET_KEY" \
            -H "Content-Type: application/json" \
            -d '{"new_dns_zone":"string","project_id":"string"}' \
            "https://api.scaleway.com/domain/v2beta1/dns-zones/{dns_zone}"
      - lang: HTTPie
        source: |-
          http PATCH "https://api.scaleway.com/domain/v2beta1/dns-zones/{dns_zone}" \
            X-Auth-Token:$SCW_SECRET_KEY \
            new_dns_zone="string" \
            project_id="string"
    delete:
      tags:
      - DNS Zones
      operationId: DeleteDNSZone
      summary: Delete a DNS zone
      description: Delete a DNS zone and all its records.
      parameters:
      - in: path
        name: dns_zone
        description: DNS zone to delete.
        required: true
        schema:
          type: string
      - in: query
        name: project_id
        description: Project ID of the DNS zone to delete.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/scaleway.domain.v2beta1.DeleteDNSZoneResponse'
      security:
      - scaleway: []
      x-codeSamples:
      - lang: cURL
        source: |-
          curl -X DELETE \
            -H "X-Auth-Token: $SCW_SECRET_KEY" \
            "https://api.scaleway.com/domain/v2beta1/dns-zones/{dns_zone}?project_id=string"
      - lang: HTTPie
        source: |-
          http DELETE "https://api.scaleway.com/domain/v2beta1/dns-zones/{dns_zone}" \
            X-Auth-Token:$SCW_SECRET_KEY \
            project_id==string
...